Rename queries->DB
